Remove SimulatedLocationManagerFactory
